+++ src/angular-generate.ts
@@ -4,13 +4,34 @@ import { getRunOutput, openUri, replaceAll } from './utilities';
 import { write, writeError, writeWN } from './logging';
 import { join } from 'path';
 import { existsSync } from 'fs';
-import { isGreaterOrEqual } from './analyzer';
+import { exists, isGreaterOrEqual, isLessOrEqual } from './analyzer';
 import { window } from 'vscode';
-import { QueueFunction } from './tip';
+import { QueueFunction, Tip, TipType } from './tip';
 import { exState } from './wn-tree-provider';
 import { npx } from './node-commands';
 import { checkAngularJson } from './rules-angular-json';
 import { getStringFrom } from './utils-strings';
+
+export async function addAngularGenerateAction(project: Project) {
+  // if (!project.isCapacitor) return;
+  if (!exists('@angular/core')) return;
+
+  project.setSubGroup('New', TipType.Add, 'Create new Angular Components, Pages and more');
+
+  const types = ['Component', 'Service', 'Module', 'Class', 'Directive'];
+  if (exists('@ionic/angular-toolkit')) {
+    types.push('Page');
+  }
+  types.forEach((item) => {
+    project.add(
+      new Tip(item, '', TipType.Angular)
+        .setQueuedAction(angularGenerate, project, item.toLowerCase())
+        .setTooltip(`Create a new Angular ${item.toLowerCase()}`)
+        .canRefreshAfter(),
+    );
+  });
+  project.clearSubgroup();
+}
 
 export async function angularGenerate(
   queueFunction: QueueFunction,
@@ -29,15 +50,13 @@ export async function angularGenerate(
   try {
     let args = '';
     if (isGreaterOrEqual('@angular/core', '15.0.0')) {
-      if (isGreaterOrEqual('@ionic/angular-toolkit', '8.1.0')) {
-        if (angularType == 'page') {
-          args += ' --standalone';
-        }
+      const isOlder = exists('@ionic/angular-toolkit') && isLessOrEqual('@ionic/angular-toolkit', '8.1.0');
+      if (angularType == 'page' && !isOlder) {
+        args += ' --standalone';
       }
-      if (isGreaterOrEqual('@ionic/angular-toolkit', '11.0.1')) {
-        if (angularType == 'component') {
-          args += ' --standalone';
-        }
+      const isOld = exists('@ionic/angular-toolkit') && isLessOrEqual('@ionic/angular-toolkit', '11.0.1');
+      if (angularType == 'component' && !isOld) {
+        args += ' --standalone';
       }
     }
     name = replaceAll(name, ' ', '-').trim();
@@ -45,7 +64,7 @@ export async function angularGenerate(
     checkAngularJson(project);
     const angularProjectName = exState.project ?? 'app';
     // eg ng generate page page-a --standalone --project=app
-    let cmd = `${npx(project)} ng generate ${angularType} ${name}${args} --project=${angularProjectName}`;
+    let cmd = `${npx(project)} ng generate ${angularType} ${name}${args}`; // --project=${angularProjectName}`;
     if (angularType == 'directive') {
       cmd += ` --skip-import`;
     }